+++ src/main/java/com/palantir/docker/compose/containers/ContainerNames.java
@@ -1,4 +1,4 @@
-package com.palantir.docker.compose;
+package com.palantir.docker.compose.containers;
 
 import static java.util.Collections.emptyList;
 import static java.util.Collections.singletonList;
