+++ src/main/java/com/palantir/docker/compose/machine/ports/Ports.java
@@ -1,4 +1,4 @@
-package com.palantir.docker.compose;
+package com.palantir.docker.compose.machine.ports;
 
 import static java.util.stream.Collectors.joining;
