+++ src/test/java/com/palantir/docker/compose/containers/logging/FileLogCollectorTest.java
@@ -1,8 +1,9 @@
-package com.palantir.docker.compose;
+package com.palantir.docker.compose.containers.logging;
 
+import static com.palantir.docker.compose.matchers.IOMatchers.fileContainingString;
+import static com.palantir.docker.compose.matchers.IOMatchers.fileWithName;
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
-
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.arrayContaining;
 import static org.hamcrest.Matchers.arrayContainingInAnyOrder;
@@ -14,15 +15,14 @@ import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import static com.palantir.docker.compose.IOMatchers.fileContainingString;
-import static com.palantir.docker.compose.IOMatchers.fileWithName;
-
+import com.palantir.docker.compose.containers.ContainerNames;
+import com.palantir.docker.compose.containers.LogCollector;
+import com.palantir.docker.compose.execution.DockerComposeExecutable;
 import java.io.File;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
-
 import org.apache.commons.io.IOUtils;
 import org.junit.Before;
 import org.junit.Rule;
